+++ Public/Get-WhisperSubtitleConfig.ps1
@@ -1,0 +1,7 @@
+function Get-WhisperSubtitleConfig {
+    [CmdletBinding()]
+    param()
+
+    . (Join-Path $PSScriptRoot '..\Private\Get-ModuleConfig.ps1')
+    Get-ModuleConfig
+}